+++ minirt_dir/set_func.c
@@ -6,13 +6,13 @@
 /*   By: namejojo <namejojo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/10 11:35:51 by namejojo          #+#    #+#             */
-/*   Updated: 2025/10/10 13:04:14 by namejojo         ###   ########.fr       */
+/*   Updated: 2025/10/11 13:11:49 by namejojo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minirt.h"
 
-t_point	set_class(float x, float y, float z)
+t_point	set_class(double x, double y, double z)
 {
 	t_point	ret;
